better output

+++ nmrfilter.py
@@ -8,24 +8,30 @@ project=sys.argv[1]
 cp = readprops(project)
 datapath=cp.get('datadir')
 
-predictionoutputfile=datapath+os.sep+project+os.sep+cp.get('predictionoutput')
-if os.path.exists(predictionoutputfile):
-	os.remove(predictionoutputfile)
-clusteringoutputfile=datapath+os.sep+project+os.sep+cp.get('clusteringoutput')
-if os.path.exists(clusteringoutputfile):
-	os.remove(clusteringoutputfile)
-louvainoutputfile=datapath+os.sep+project+os.sep+cp.get('louvainoutput')
-if os.path.exists(louvainoutputfile):
-	os.remove(louvainoutputfile)
-predictionoutputfile=datapath+os.sep+project+os.sep+cp.get('predictionoutput')+'hsqc'
-if os.path.exists(predictionoutputfile):
-	os.remove(predictionoutputfile)
-predictionoutputfile=datapath+os.sep+project+os.sep+cp.get('predictionoutput')+'hmbc'
-if os.path.exists(predictionoutputfile):
-	os.remove(predictionoutputfile)
-predictionoutputfile=datapath+os.sep+project+os.sep+cp.get('predictionoutput')+'hsqctocsy'
-if os.path.exists(predictionoutputfile):
-	os.remove(predictionoutputfile)
+if not os.path.exists(datapath+os.sep+project):
+	print("There is no directory "+datapath+os.sep+project+" - please check!")
+
+if os.path.exists(datapath+os.sep+project+os.sep+"result"):
+	predictionoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('predictionoutput')
+	if os.path.exists(predictionoutputfile):
+		os.remove(predictionoutputfile)
+	clusteringoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('clusteringoutput')
+	if os.path.exists(clusteringoutputfile):
+		os.remove(clusteringoutputfile)
+	louvainoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('louvainoutput')
+	if os.path.exists(louvainoutputfile):
+		os.remove(louvainoutputfile)
+	predictionoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('predictionoutput')+'hsqc'
+	if os.path.exists(predictionoutputfile):
+		os.remove(predictionoutputfile)
+	predictionoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('predictionoutput')+'hmbc'
+	if os.path.exists(predictionoutputfile):
+		os.remove(predictionoutputfile)
+	predictionoutputfile=datapath+os.sep+project+os.sep+'result'+os.sep+cp.get('predictionoutput')+'hsqctocsy'
+	if os.path.exists(predictionoutputfile):
+		os.remove(predictionoutputfile)
+else:
+	os.mkdir(datapath+os.sep+project+os.sep+"result")
 
 if os.path.exists(datapath+os.sep+project+os.sep+"plots"):
 	for f in os.listdir(datapath+os.sep+project+os.sep+"plots"):
